Add some functions in init.wrapper
Now it resizes rootfs. Resizing was originally done by systemd-r-
size service. But now read-only option is going to be applied, so
it can not be done as systemd service. So init.wrapper took the
job. Initrd will take the resizing job later.

Also now it uses both of 'partlabel' and 'label' to find the disk.
Before this commit it did not find the disk by 'partlabel', so it
has not been working.

Change-Id: Ibb56a459e614b3f83651806c246fe3b877e5441d